Updates to text, document picker and inspectors:
Replaced the OUIInspectorDidDismiss notification with a delegate method -inspectorDidDismiss:.  This allows the core text editor to use its own OUIInspector without the OUIDocument inappropriately responding to the notifications of this other OUIInspector.

Tapping the document title on the picker will now allow you to rename the document, as in iWork.

100% fewer skeletons in the rectangular loupe.

Keep UITextInputStringTokenizer's textInput in our own ivar instead of referencing the superclass ivar.

RTF underline support, since our sample doc has it