[r136540] Factored out the drag-hysteresis tracking into a new gesture recognizer subclass, OUILongPressGestureRecognizer.
Just add an "O" to the beginning of your gesture recognizer class name, and you get hysteresis tracking for free.  There is also a method -resetHysteresis which causes the tracking to start over from the current location of the finger(s).  This is useful for nudging -- if the first finger is slowly drifting while the user performs a nudge swipe with a second finger, we don't want that first finger's movement to interfere with the precise nudging (so we reset the hysteresis with every nudge).  I converted some of the GesturePlayground classes to use this new abstracted hysteresis tracking.

This basically completes the functionality of the built-in UILongPressGestureRecognizer when used for dragging/resizing/panning.  In case I haven't said this before, there are a lot of reasons to use LongPress GRs instead of Pan GRs -- even when you don't care about using the built-in delay.  One reason is that they will always "recognize" before the UIScrollView's Pan GR kicks in.  Another reason is that a LongPress GR with a delay of about 0.02 seconds leaves enough time for a second finger to hit the screen, cleanly allowing two-finger gestures to occur.

(The default UILongPressGestureRecognizer does in fact utilize hysteresis behind the scenes, but it only takes effect before the delay timer finishes.  i.e. if your finger doesn't move more than 10 pixels before the timer fires, it's considered a long press.  But any movement *after* that causes a "changed" event regardless.  The hysteresis tracking I added gives us a way to ask whether the finger has really moved far enough to start performing whatever action the GR is supposed to do.  if (![gestureRecognizer overcameHysteresis]) return;  )