[r136861] fix for display mirroring:
Don't turn on geometryFlipped on our layer. This was left over from when we implemnted -drawRect:, most likely and had no effect other than to mean we needed to pass a different argument to the shadow view set up. Now that we set the layer's content directly in -displayLayer: we don't need it.

This avoids bugs in -[CALayer renderInContext:], which we need to use for the updated screen mirroring support