[r138084] Improve inspector resizing support
The inspector window wasn't resizing as needed when the column type was changed. This is due to the general inspector rearrangement in OO4, which revealed that there wasn't a mechanism in place to prompt parent inspectors from laying out again to accommodate size changes in descendant inspectors.

Added a new method, -inspectorDidResize:, on both OIInspector and OIInspectorController. When an inspector (in this case, the column type inspector) changes its size programmatically, it informs its inspector controller via -inspectorDidResize:. The controller simply passes the message on to the root inspector (if that is not the same as the one that resized).

OIInspector's implementation of -inspectorDidResize: doesn't do anything except assert that it never gets called (because this method shouldn't be called on the inspector that initiated the notification). OISectionedInspector and OITabbedInspector notify any appropriate child inspectors, and then relayout to accommodate the size changes.

Modified -setInspectorController: on OITabbedInspector and OISectionedInspector to also set the inspector controller on any tabs/sections, so that they know the inspector controller and can notify it of height changes.

Modified -[OOColumnTypeInspector _redisplay] to call -[OIInspectorController inspectorDidResize:] instead of calling -updateExpandedness:, which doesn't have the context that we have available to us. Also, modified it to assume (and assert) that the inspector view is inside a flipped view, since it now is, with the inspector rearrangement. And be more defensive about asking for the controlsView frame, since the analyzer doesn't like us calling a struct-returning method on an object that could be nil.