[r137225] Make text stick to the top of OUIEditableFrame instead of the bottom.
In field editors, if the height of the editor is taller than the text (perhaps it needs to span a minimum height), the text needs to stick to the origin like in NSTextView and friends.

There are still remaining issues (as there were before) the transition between using OUIEditableFrame and OUITextLayout with multiple lines. Something about how we are (or aren't) positioning/pixel-snapping in the two cases differs, causing jumpiness in the apparent text origin and jumpiness in line spacing as new lines are added. Wim will likely need to look at what is going on here.

OUITextLayout:
- Factored out a couple new functions, OUITextLayoutOrigin() and OUITextLayoutDrawFrame() and use them in both the OUITextLayout and OUIEditableFrame drawing paths.
- OUITextLayoutOrigin() calculates the layout origin like OUIEditableFrame used to, except that it pins the text to the top.
- OUITextLayoutDrawFrame() takes the bounds to draw in and the layout origin to use and does what OUIEditableFrame used to, but now OUITextLayout can use it.

OUIEditableFrame:
- Use the new functions in OUITextLayout to get the layout origin (pinning the text to the top now) and draw.
- Set our content mode to UIViewContentModeTopLeft so that when animating resizing we don't stretch our text.

OmniUI's TextEditor example:
- Adjust the layout of the view such that the editable text takes up the full width of the scroll view and at least its height, but possible more if the text needs it.
- Fix the effective page text width to something like 8 1/2 inches.
- Zooming likely needs work, if it works at all.