Added support for controlling the output encoding and directory. Rather than using the internal Xcode "copystrings" tool we were using, this is now a suitable replacement.
Add 'binary' as an option for spitting out strings.
This emits a bplist, like the copystrings tool did.

Handle inputs that are already in bplist format. In this case we require the output format to be "binary" too.
This happens when doing an iOS build when the frameworks export a strings file with "binary" and then the app copies it in.

Added missing super dealloc.