Added some scrolling behavior for <https://github.com/omnigroup/OmniGroup/issues/37>
OUIEditableFrame intentaionlly doesn't scroll when typing/selection changes happen since it doens't have enough information to do so. In simple examples like this, it might be able to, but in more complex examples where the editable frame is a subview of some possibly nested set of scroll views, it cannot.