[r161389] <bug:///77585> (document layout animation stutters when creating a new document)
OFSDocumentStore scanItemsWithCompletionHandler: scan-finished-block has a couple of places where the scope is asked for its container URL. This uses -[NSFileManager URLForUbiquityContainerIdentifier:] and I have found out that that can be slow. And, when this is called on each file in the document store, the app can slow down a lot.
Therefore, I added a couple of convenience functions to generate a NSDictionary cache of the store scope containers and is used within these file loops. Note that a couple of the locations are actually assertions and have been ifdef'd appropriately. In addition, the assertion in OFSDocumentStoreFileItem _updateWithMetadataItem: was redundant and expensive so I commented it out for now.