[r161449] <bug:///77920> (Tapping on two file names at the same time blocks interaction with the previews)
It seems that, at least in iOS 5.1, that UIGestureRecognizer queues its actions asynchronously.

This means that two recognizers can accept and queue actions before the first's action has fired.

Rather than asserting that we aren't already in the middle of renaming when a rename is started, actually check and just bail if so.